changes in the feed when posting and commenting
